improve build process
