+++ .gitignore
@@ -1,10 +1,7 @@
 .idea
 *.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
 node_modules/
-dist/
+lib/
+bin/
 .npm
 .env
