+++ packages/core/package.json
@@ -2,16 +2,23 @@
   "name": "@exogee/kew",
   "version": "0.4.0",
   "description": "A typed task queue with a simple API",
-  "main": "dist/index.js",
-  "source": "src/index.ts",
+  "exports": {
+    ".": {
+      "require": "./lib/index.cjs",
+      "import": "./lib/index.mjs"
+    }
+  },
+  "source": "./src/index.ts",
+  "types": "./types/index.d.ts",
   "scripts": {
-    "build": "esbuild src/index.ts --outfile=dist/index.js --bundle --format=esm"
+    "build": "pnpm build:esm && pnpm build:cjs && pnpm build:types",
+    "build:esm": "esbuild src/index.ts --outfile=lib/index.mjs --bundle --format=esm",
+    "build:cjs": "esbuild src/index.ts --outfile=lib/index.cjs --bundle --format=cjs",
+    "build:types": "tsc src/**/*.ts --outDir lib --declaration --emitDeclarationOnly"
   },
   "license": "MIT",
-  "dependencies": {
-    "typescript": "4.3.5"
-  },
   "devDependencies": {
+    "typescript": "4.3.5",
     "esbuild": "0.15.13",
     "esbuild-register": "3.4.1"
   }
